==> 01-get-data-by-metro.ipynb <==
# Get Data by Metro Area

## 1.0. Load Packages

# Import packages to use later
import pandas as pd
import nltk
import time
import json

## 1.1. Investigate Cities
### We'll have 2 files that we're working with in this file. 1 is businesses; the second is the reviews from these businesses.

businesses = pd.read_csv(r"YelpChallengeWMetros_Clean.csv", encoding='latin-1')
businesses = businesses.rename(columns={'business': 'business_id'})

d = dict(businesses['metroarea'].value_counts())

for city, count in sorted(d.items(), key=lambda x: x[1], reverse=True):
    if count > 1000:
        print(city, count)

businesses.head(3)

## 1.2. Dataset for Each Metro

#### Due to Yelp review dataset size, the remainder of this notebook (Section 1.2.) is unused for this demonstration. If you were to continue this work beyond this demo, we have included some advice on processing the Yelp dataset from scratch in the README.

def create_metro_csvs():
    # These cities contain almost all reviews in the dataset (removing outlier entries)
    metros = [city for city, count in sorted(d.items(), key=lambda x: x[1], reverse=True)]

    # Initialize empty dataframe for each city
    metro_reviews = {metro:pd.DataFrame() for metro in metros}

    # Dictionary of city IDs to set of business IDs in that city
    metro_businesses = {metro:set(businesses[businesses['metroarea'] == metro]['business_id'].tolist()) for metro in metros}

    num_total_reviews = 0

    with pd.read_csv(r"yelp_academic_dataset_reviews_Boulder.csv", chunksize=100000) as reader:
        for count, chunk in enumerate(reader):
            #####
            chunk['datetime'] = pd.to_datetime(chunk['date'])
            chunk = chunk[chunk['datetime'] >= pd.Timestamp('2010-01-01')]
            chunk = chunk[chunk['datetime'] < pd.Timestamp('2020-01-01')]
            #print(chunk.head(5))
            #####
            num_total_reviews += len(chunk)
            for metro in metros:
                vals = chunk[chunk.business_id.isin(metro_businesses[metro])]
                metro_reviews[metro] = pd.concat([metro_reviews[metro], vals], axis=0)
            if count % 10 == 0:
                print("Chunk #", count)
            #break
        print("End: Chunk #", count)

    print(num_total_reviews)

# we're commenting out this line because we only have reviews for one metro - Boulder
# for metro in metros:
#     #metro_reviews[metro].to_csv("small_reviews/yelp_academic_dataset_reviews_" + metro + ".csv")
#     metro_reviews[metro].to_csv("small_reviews_urbcomp/yelp_academic_dataset_reviews_" + metro + ".csv")

## Checking Results

def check_results():
    print(metros)
    lengths = [len(metro_reviews[metro]) for metro in metros]
    print(lengths)
    print(sum(lengths))
    print(num_total_reviews)

#check_results()
