==> 04-output-files.ipynb <==
## 4.0. Load Packages

import pandas as pd
import matplotlib.pyplot as plt
import json
import numpy as np

## 4.1. Load Business Category Data

# Get top-level business categories
with open("categories.json") as json_file:
    categories = json.load(json_file)

# Dictionary of category keys to the full strings that appear in the dataset
category_dict = {category["alias"]:category["title"] for category in categories}
category_dict_reverse = {category["title"]:category["alias"] for category in categories}

# Categories that we are interested in
#categories_of_interest = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
categories_of_interest_raw = [category["alias"] for category in categories if category["parents"] == []]
categories_of_interest = [category_dict[cat] for cat in categories_of_interest_raw]

print(categories_of_interest)
print(categories_of_interest_raw)

## 4.2. Business Data

### Load Dataset

businesses = pd.read_csv("YelpChallengeWMetros_Clean.csv", encoding='latin-1')
# businesses.head(3)

### Process Categories

# Create indicator columns for each business category (a business may belong to multiple categories)
# such as: is_restaurants or is_shopping
businesses_categories = businesses.copy()

for category_str in categories_of_interest:
    colname = "is_" + category_dict_reverse[category_str]
    businesses_categories[colname] = businesses_categories["categories"].str.contains(category_str).fillna(False)
businesses_categories.head(3)

## 4.3. Counts Data (all metro areas combined) and num_businesses

### counts_df

#Takes the previously calculated metros_counts_final_v2.csv files and concatenates them into 1 file for analysis below
metros = ['Boston', 'Portland', 'Austin', 'Orlando', 'Atlanta', 'Vancouver', 'Columbus', 'Boulder']

counts_df = pd.DataFrame()


for metro in metros:

    suffix = "_v2.csv"
        
    counts_df_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_final" + suffix)
    counts_df_metro["metroarea"] = metro
    counts_df = pd.concat([counts_df,counts_df_metro])
    
counts_df.to_csv("final_output_urbcomp/all_counts_final_v2.csv", index=False)

#### num_businesses

#Outputs a dataframe with 1 metro per row and a count of number of businesses and times a word (column) shows up in business reviews in that metro

categories_of_interest_small = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
counts_df_mod_noprofs = counts_df.copy()
for col in ["is_"+col for col in categories_of_interest_small if col != "restaurants"]:
    counts_df_mod_noprofs[col] = np.logical_and(counts_df_mod_noprofs[col] == True, counts_df_mod_noprofs["is_restaurants"] == False)
cols = ["is_"+col for col in categories_of_interest_small]

df_rows = []
# categories_of_interest_small = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']

cols = [counts_df.columns[0]] + list(counts_df.columns[3:45])

for category in categories_of_interest_small:
    df_row = [category]
    q = "is_" + category
    counts_category = counts_df_mod_noprofs.query(q)
    num_bus_in_cat = len(counts_category)
    counts_df_temp = counts_category[cols[1:]]
    df_row.append(num_bus_in_cat)
    df_row.extend(np.count_nonzero(counts_df_temp, axis=0))
    df_rows.append(df_row)
    
num_businesses_df = pd.DataFrame(df_rows,columns = ["category", "num_businesses"] + cols[1:])
    
num_businesses_df.to_csv("final_output_urbcomp/num_businesses.csv", index=False)
num_businesses_df.head(3)

## 4.4. counts_df_norm and table

#### Initial Data Prep

#Takes the previously defined metro_counts_norm_final_v2.csv files and appends them all together

metros = ['Boston', 'Portland', 'Austin', 'Orlando', 'Atlanta', 'Vancouver', 'Columbus', 'Boulder']

counts_df_norm = pd.DataFrame()

for metro in metros:
    suffix = "_v2.csv"
        
    counts_df_norm_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_norm_final" + suffix)
    counts_df_norm_metro["metroarea"] = metro
    counts_df_norm = pd.concat([counts_df_norm,counts_df_norm_metro])
    
counts_df_norm.head(3)

#### Table

metros = ['Boston', 'Portland', 'Austin', 'Orlando', 'Atlanta', 'Vancouver', 'Columbus', 'Boulder']
metros = sorted(metros)
cols = list(counts_df_norm.columns[3:41])

#Outputs a full table of normalized counts of relationship words per metro and overall, to a text document

df = pd.DataFrame(counts_df_norm[cols].mean().sort_values(ascending=False))

for metro in metros:
    counts_df_norm_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_norm_final_v2.csv")
    df[metro] = pd.DataFrame(counts_df_norm_metro[cols].mean().sort_values(ascending=False))
df = df.reset_index()
df = df.rename(columns={"index": "Word", 0: "Total"})
df.head(3)

df.to_latex(buf="output_urbcomp/table.txt", float_format="%.2f", index=False)

#### counts_df_norm file

categories_of_interest = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
#creates logic statement for the following query statement, to get those rows where the columns selected are TRUE in at least 1 of the columns
q = " or ".join(["is_"+cat for cat in categories_of_interest])
print(q)

#If a business is registered as is_shopping=True and is_restaurants=True, this sets is_shopping=False
#Basically makes it so a restaurant is not able to be counted as another category

final_table = counts_df_norm.query(q).copy()

for col in ["is_"+col for col in categories_of_interest if col != "restaurants"]:
    final_table[col] = np.logical_and(final_table[col] == True, final_table["is_restaurants"] == False)

final_table.to_csv("final_output_urbcomp/all_counts_norm_final_v2.csv", index=False)

#### counts_df_norm_formaps file

#Takes just the columns from final_table that are used for mapping and outputs them into a separate dataset
cols_of_interest = ['business_id', 'num_reviews', 'num_relationship_words',
                    'family', 'romantic', 'friendship', 'professional',
                    'name', 'categories', 'stars', 'review_count',
                    'address', 'city', 'state', 'longitude', 'latitude', 'metroarea']

cols_of_interest.extend(["is_"+cat for cat in categories_of_interest])

counts_df_norm_fewcols = final_table[cols_of_interest]
counts_df_norm_fewcols.to_csv("final_output_urbcomp/all_counts_norm_fewcols_formaps.csv", index=False)
